Refactor tests for data
