+++ tests/test_data/test_timeseries.py
@@ -1,5 +1,4 @@
 from copy import deepcopy
-import itertools
 import pickle
 from typing import Dict
 
@@ -9,125 +8,58 @@ import pytest
 from sklearn.preprocessing import StandardScaler
 import torch
 
-from pytorch_forecasting.data import (
-    EncoderNormalizer,
-    GroupNormalizer,
-    NaNLabelEncoder,
-    TimeSeriesDataSet,
-    TimeSynchronizedBatchSampler,
-)
+from pytorch_forecasting.data import EncoderNormalizer, GroupNormalizer, NaNLabelEncoder, TimeSeriesDataSet
 from pytorch_forecasting.data.encoders import MultiNormalizer, TorchNormalizer
 from pytorch_forecasting.data.examples import get_stallion_data
 from pytorch_forecasting.data.timeseries import _find_end_indices
 from pytorch_forecasting.utils import to_list
 
-torch.manual_seed(23)
+
+def test_find_end_indices():
+    diffs = np.array([1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1])
+    max_lengths = np.array([4, 4, 4, 4, 4, 4, 4, 4, 3, 2, 1, 4, 4, 4, 4, 4, 4, 4, 4, 3, 2, 1])
+
+    ends, missings = _find_end_indices(diffs, max_lengths, min_length=3)
+    ends_test = np.array([3, 4, 4, 5, 6, 8, 9, 10, 10, 10, 10, 14, 15, 15, 16, 17, 19, 20, 21, 21, 21, 21])
+    missings_test = np.array([[0, 2], [5, 7], [11, 13], [16, 18]])
+    np.testing.assert_array_equal(ends, ends_test)
+    np.testing.assert_array_equal(missings, missings_test)
 
 
-@pytest.mark.parametrize(
-    "data,allow_nan",
-    itertools.product(
-        [
-            (np.array([2, 3, 4]), np.array([1, 2, 3, 5, np.nan])),
-            (np.array(["a", "b", "c"]), np.array(["q", "a", "nan"])),
-        ],
-        [True, False],
-    ),
-)
-def test_NaNLabelEncoder(data, allow_nan):
-    fit_data, transform_data = data
-    encoder = NaNLabelEncoder(warn=False, add_nan=allow_nan)
-    encoder.fit(fit_data)
-    assert np.array_equal(
-        encoder.inverse_transform(encoder.transform(fit_data)), fit_data
-    ), "Inverse transform should reverse transform"
-    if not allow_nan:
-        with pytest.raises(KeyError):
-            encoder.transform(transform_data)
-    else:
-        assert encoder.transform(transform_data)[0] == 0, "First value should be translated to 0 if nan"
-        assert encoder.transform(transform_data)[-1] == 0, "Last value should be translated to 0 if nan"
-        assert encoder.transform(fit_data)[0] > 0, "First value should not be 0 if not nan"
+def test_raise_short_encoder_length(test_data):
+    with pytest.warns(UserWarning):
+        test_data = test_data[lambda x: ~((x.agency == "Agency_22") & (x.sku == "SKU_01") & (x.time_idx > 3))]
+        TimeSeriesDataSet(
+            test_data,
+            time_idx="time_idx",
+            target="volume",
+            group_ids=["agency", "sku"],
+            max_encoder_length=5,
+            max_prediction_length=2,
+            min_prediction_length=1,
+            min_encoder_length=5,
+        )
 
 
-def test_NaNLabelEncoder_add():
-    encoder = NaNLabelEncoder(add_nan=False)
-    encoder.fit(np.array(["a", "b", "c"]))
-    encoder2 = deepcopy(encoder)
-    encoder2.fit(np.array(["d"]))
-    assert encoder2.transform(np.array(["a"]))[0] == 0, "a must be encoded as 0"
-    assert encoder2.transform(np.array(["d"]))[0] == 3, "d must be encoded as 3"
-
-
-@pytest.mark.parametrize(
-    "kwargs",
-    [
-        dict(method="robust"),
-        dict(transformation="log"),
-        dict(transformation="softplus"),
-        dict(transformation="log1p"),
-        dict(transformation="relu"),
-        dict(center=False),
-    ],
-)
-def test_EncoderNormalizer(kwargs):
-    data = torch.rand(100)
-    defaults = dict(method="standard", center=True)
-    defaults.update(kwargs)
-    kwargs = defaults
-    normalizer = EncoderNormalizer(**kwargs)
-    if kwargs.get("transformation") in ["relu", "softplus"]:
-        data = data - 0.5
-
-    if kwargs.get("transformation") in ["relu", "softplus", "log1p"]:
-        assert (
-            normalizer.inverse_transform(normalizer.fit_transform(data)) >= 0
-        ).all(), "Inverse transform should yield only positive values"
-    else:
-        assert torch.isclose(
-            normalizer.inverse_transform(normalizer.fit_transform(data)), data, atol=1e-5
-        ).all(), "Inverse transform should reverse transform"
-
-
-@pytest.mark.parametrize(
-    "kwargs,groups",
-    itertools.product(
-        [
-            dict(method="robust"),
-            dict(transformation="log"),
-            dict(transformation="relu"),
-            dict(center=False),
-            dict(transformation="log1p"),
-            dict(transformation="softplus"),
-            dict(scale_by_group=True),
-        ],
-        [[], ["a"]],
-    ),
-)
-def test_GroupNormalizer(kwargs, groups):
-    data = pd.DataFrame(dict(a=[1, 1, 2, 2, 3], b=[1.1, 1.1, 1.0, 5.0, 1.1]))
-    defaults = dict(method="standard", transformation=None, center=True, scale_by_group=False)
-    defaults.update(kwargs)
-    kwargs = defaults
-    kwargs["groups"] = groups
-    kwargs["scale_by_group"] = kwargs["scale_by_group"] and len(kwargs["groups"]) > 0
-
-    if kwargs.get("transformation") in ["relu", "softplus"]:
-        data.b = data.b - 2.0
-    normalizer = GroupNormalizer(**kwargs)
-    encoded = normalizer.fit_transform(data["b"], data)
-
-    test_data = dict(
-        prediction=torch.tensor([encoded[0]]),
-        target_scale=torch.tensor(normalizer.get_parameters([1])).unsqueeze(0),
+def test_categorical_target(test_data):
+    dataset = TimeSeriesDataSet(
+        test_data,
+        time_idx="time_idx",
+        target="agency",
+        group_ids=["agency", "sku"],
+        max_encoder_length=5,
+        max_prediction_length=2,
+        min_prediction_length=1,
+        min_encoder_length=1,
     )
 
-    if kwargs.get("transformation") in ["relu", "softplus", "log1p"]:
-        assert (normalizer(test_data) >= 0).all(), "Inverse transform should yield only positive values"
-    else:
-        assert torch.isclose(
-            normalizer(test_data), torch.tensor(data.b.iloc[0]), atol=1e-5
-        ).all(), "Inverse transform should reverse transform"
+    _, y = next(iter(dataset.to_dataloader()))
+    assert y[0].dtype is torch.long, "target must be of type long"
+
+
+def test_pickle(test_dataset):
+    pickle.dumps(test_dataset)
+    pickle.dumps(test_dataset.to_dataloader())
 
 
 def check_dataloader_output(dataset: TimeSeriesDataSet, out: Dict[str, torch.Tensor]):
@@ -339,80 +271,6 @@ def test_overwrite_values(test_dataset, value, variable, target):
         changed = torch.isclose(outputs[0][name], control_outputs[0][name]).all()
         assert changed, f"Output {name} should be reset"
     assert torch.isclose(outputs[1][0], control_outputs[1][0]).all(), "Target should be reset"
-
-
-@pytest.mark.parametrize(
-    "drop_last,shuffle,as_string,batch_size",
-    [
-        (True, True, True, 64),
-        (False, False, False, 64),
-        (True, False, False, 1000),
-    ],
-)
-def test_TimeSynchronizedBatchSampler(test_dataset, shuffle, drop_last, as_string, batch_size):
-    if as_string:
-        dataloader = test_dataset.to_dataloader(
-            batch_sampler="synchronized", shuffle=shuffle, drop_last=drop_last, batch_size=batch_size
-        )
-    else:
-        sampler = TimeSynchronizedBatchSampler(
-            data_source=test_dataset, shuffle=shuffle, drop_last=drop_last, batch_size=batch_size
-        )
-        dataloader = test_dataset.to_dataloader(batch_sampler=sampler)
-
-    time_idx_pos = test_dataset.reals.index("time_idx")
-    for x, _ in iter(dataloader):  # check all samples
-        time_idx_of_first_prediction = x["decoder_cont"][:, 0, time_idx_pos]
-        assert torch.isclose(
-            time_idx_of_first_prediction, time_idx_of_first_prediction[0]
-        ).all(), "Time index should be the same for the first prediction"
-
-
-def test_find_end_indices():
-    diffs = np.array([1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1])
-    max_lengths = np.array([4, 4, 4, 4, 4, 4, 4, 4, 3, 2, 1, 4, 4, 4, 4, 4, 4, 4, 4, 3, 2, 1])
-
-    ends, missings = _find_end_indices(diffs, max_lengths, min_length=3)
-    ends_test = np.array([3, 4, 4, 5, 6, 8, 9, 10, 10, 10, 10, 14, 15, 15, 16, 17, 19, 20, 21, 21, 21, 21])
-    missings_test = np.array([[0, 2], [5, 7], [11, 13], [16, 18]])
-    np.testing.assert_array_equal(ends, ends_test)
-    np.testing.assert_array_equal(missings, missings_test)
-
-
-def test_raise_short_encoder_length(test_data):
-    with pytest.warns(UserWarning):
-        test_data = test_data[lambda x: ~((x.agency == "Agency_22") & (x.sku == "SKU_01") & (x.time_idx > 3))]
-        TimeSeriesDataSet(
-            test_data,
-            time_idx="time_idx",
-            target="volume",
-            group_ids=["agency", "sku"],
-            max_encoder_length=5,
-            max_prediction_length=2,
-            min_prediction_length=1,
-            min_encoder_length=5,
-        )
-
-
-def test_categorical_target(test_data):
-    dataset = TimeSeriesDataSet(
-        test_data,
-        time_idx="time_idx",
-        target="agency",
-        group_ids=["agency", "sku"],
-        max_encoder_length=5,
-        max_prediction_length=2,
-        min_prediction_length=1,
-        min_encoder_length=1,
-    )
-
-    _, y = next(iter(dataset.to_dataloader()))
-    assert y[0].dtype is torch.long, "target must be of type long"
-
-
-def test_pickle(test_dataset):
-    pickle.dumps(test_dataset)
-    pickle.dumps(test_dataset.to_dataloader())
 
 
 @pytest.mark.parametrize(
