+++ tests/test_models/conftest.py
@@ -6,6 +6,8 @@ from pytorch_forecasting import TimeSeriesDataSet
 from pytorch_forecasting.data import EncoderNormalizer, GroupNormalizer, NaNLabelEncoder
 from pytorch_forecasting.data.examples import generate_ar_data, get_stallion_data
 
+torch.manual_seed(23)
+
 
 @pytest.fixture
 def gpus():
